// src/libs/shared/config/app-name-config.enum.ts
export enum AppServiceNameConfig {
    Id = 'id',
    Restaurants = 'restaurants',
    Suppliers = 'suppliers',
    Membership = 'membership',
    Products = 'products',
    Orders = 'orders',
    ProductsCategoires = 'products-categories',
    FilesCreators = 'files-creators',
    FilesManagers = 'files-managers',
    ShoppingCarts = 'shopping-carts',
    Company = 'company',
}